refactor: rename nodes.ts-> node.ts
match current convention of singular file names
