+++ web/src/modules/topic/components/Diagram/Diagram.tsx
@@ -15,7 +15,7 @@ import { useViewportUpdater } from "../../hooks/flowHooks";
 import { closeClaimDiagram, connectNodes, deselectNodes } from "../../store/actions";
 import { useFilteredDiagram } from "../../store/store";
 import { type Edge, type Node } from "../../utils/diagram";
-import { type NodeType } from "../../utils/nodes";
+import { type NodeType } from "../../utils/node";
 import { FlowNode } from "../Node/FlowNode";
 import { ScoreEdge } from "../ScoreEdge/ScoreEdge";
 import { PositionedCloseButton, StyledReactFlow } from "./Diagram.styles";
