+++ web/src/modules/topic/components/Indicator/Indicator.styles.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import { Button } from "@mui/material";
 
-import { indicatorLength } from "../../utils/nodes";
+import { indicatorLength } from "../../utils/node";
 
 export const StyledButton = styled(Button)`
   width: ${indicatorLength}px;
