+++ web/src/modules/topic/components/Node/AddNodeButtonGroup.tsx
@@ -3,7 +3,7 @@ import { ButtonGroup } from "@mui/material";
 import { type RelationDirection } from "../../utils/diagram";
 import { addableRelationsFrom } from "../../utils/edge";
 import { Orientation } from "../../utils/layout";
-import { NodeType } from "../../utils/nodes";
+import { NodeType } from "../../utils/node";
 import { AddNodeButton } from "../Node/AddNodeButton";
 
 interface Props {
